Fix Temp/RH display: divide by 10 for decimal values

+++ Models/SensorData.cs
@@ -9,6 +9,6 @@ public class SensorData
     public int CO { get; set; }
     public int VOC1 { get; set; }
     public int VOC2 { get; set; }
-    public int Temp { get; set; }
-    public int RH { get; set; }
+    public decimal Temp { get; set; }  // Stored as int*10, display as decimal
+    public decimal RH { get; set; }    // Stored as int*10, display as decimal
 }
